Specify that swarm 1.1 is faster
Torbjørn tests on a set of 312,503 unique amplicons with a length
average of 382 nucleotides show a 19% speed increase with the new
swarm (with d = 1). Using the uclust-like output option limits that
speed up a bit when d = 1 (17% instead of 19%), but much more when d =
7 (7% instead of 18%). The optimizations introduced in swarm 1.1
reduce the number of pairwise comparisons to do by using slightly more
memory. The gain should be noticeable for most datasets, but will vary
with the swarm parameters used, and the size and internal structure of
the dataset.