Ugly bug in the hash function
In the bash code we use to dereplicate the amplicons, I was using a
here-string '<<<' to pass the sequence to the sha1sum command. The
objective was to avoid the use of a pipe in the main loop to speed up
the script (a pipe creates a sub-process, which is a slow operation).

Unfortunately, here-strings automatically append a newline character
to the string. That newline character modifies the string, and hence
results in a different hash digest.

I spotted the bug while I was trying to use the python 'hashlib'
module and couldn't get matching results with the bash script. With
the new bash script, now passing the string with 'printf' through a
pipe, the hash digests obtained are the same than the ones obtained
with python.