modernize by using trailing return type
This is a big change in terms of style, but it is recommended by C++
leaders. As it is, swarm's code does not make use of the advantages of
using trailing returns, but it might in future versions.

This invasive change has been validated by clang-tidy: code logic is
not affected. As usual, I also ran all our unit tests and tests on
real files.