use "cluster" throughout
The traditional term "OTU" is negatively charged nowadays. The ASV vs
OTU debate is creating confusion in the community and some users now
think that all methods producing "OTUs" use a fixed clustering
threshold (i.e. 97%-similarity) and are inherently bad.

Of course, this is not the case and there are several methods
published before the ASV term was coined that produce ASV-like
clusters, swarm included.

To avoid that confusion, swarm's manual now only uses the generic term
"cluster".