generate ascii array at runtime
computation time increases a bit (by a couple hundred milliseconds
maybe), but I believe the function makes the array easier to
understand.

It also removes a warning regarding code duplication (lines of '-1'
were seen as duplicated code).

With C++14 this function is declared as constexpr, the array is
created at compile time, and the function code is not included in the
executable. Hopefully, we will move from C++11 to C++14 in the future,
and we will recoup that performance loss.