add a just-in-time hint for ill-formatted options
This is an attempt to address GitHub issue #173

The current internal test for command line arguments outputs a generic
error message ("invalid numeric arguments") when an option is
ill-formatted.

The error can be caused by a single-dash for a long option name, or a
missing space, or by an invalid numerical value. The latter case
concerns options expecting numerical values and is covered by specific
tests. Options expecting a name for an output file are not
impacted. Specific tests are also present for some options that do not
expect arguments (i.e. hvxnfr).

To summarize, it seems that using a single-dash for a long option name
or forgetting a space between an argument and an option are the most
likely causes for this error message. This extended error message
should make it easier for users to find what is wrong with their
command line.