bug fix: bloomflex std::bad_alloc for large (2^32?) allocations
when using the fastidious option, swarm crashes when it tries to
allocate the bloomfilter (at least for allocations of 5 GB, could also
be the case for smaller allocations).

Our current tests are not able to trigger this bug.

I've introduced the bug on January 28th while replacing malloc with
new and delete. Sadly, the bug is present in main branch and in the
v3.1.1 release, so I think an urgent v3.1.2 release is necessary.

Other replacements of malloc made in January may have introduced
so-far undetected bugs. I will do my best to test large datasets
before future releases.