'static constexpr' local variables (function scope)
In the context of a function scope 'constexpr' values are stack
values. Marking them as 'static' forces initialization at compile
time, and are not placed in the local stack at run time. In the global
scope, 'constexpr' values are global, and should not be marked as
'static' (but could be marked as 'inline').