replace range-for loop with algorithm for_each and a lambda
The range-for loop compiled to a much smaller number of instructions
than the naked C-loops (from 242 instructions to 185, including code
to output the matrix to stdout), and seems to be using vectorized (SSE
at least).

Moving to a for_each and a lambda yields a slightly smaller number of
instructions (from 185 instructions to 184), but it is a first step
toward a more functional style of programming.

Above compilation tests were done with the latest GCC (trunk) and
targetting a x86-64 architecture.