double the maximal number of threads (256 to 512)
In 2023, the new AMD EPYC 9654 can have up to 96 cores and 192
threads. With a dual-CPU set up, it is possible to reach 192 cores and
384 threads. Similarly, there are high-thread-count ARM computers
available. In that context, capping swarm to a maximum of 256 threads
could be limiting for some users. I propose to set a new limit at 512
threads, for now.